Update main.css
Here is another option: shrink your buttons by 40px, then add the height back with a 20px margin on top and bottom. I left your big shadow on this one. I noticed it breaks one of your buttons where the text is so long that it extends beyond the button.