Update main.css
And here is a third option. What I've done here is remove the color fill (and the Army removed the connection to your CDN, so, good job, Army, I didn't need icon fonts for more than ten minutes anyway) and add a transition. This way the piece feels more natural as it raises off the page. It shouldn't snap instantly. Our minds don't understand that movement.

Also, I pushed some of the hover styles into the default style. Hovers are for changing things. That's why, before, your width would jump from whatever it was to 100%.

Transitions work the same way. You put the transition on the default style, and it uses that information to move between changes. In this case, your box shadow is changing between regular and :hover states. The transition slows this down to make awesome happen.

Enjoy!